Update README.md
This is a crypto comparator which utilizes two crypto APIs to retrieve the data for crypto assets and compare them. First crypto API is used to display data in react table with images and corresponding current crypto data. Second API is used to retrieve more in-depth data and compare selected crypto assets. Statistics are displayed on the final page along with graphs and recommendation which asset is the best choice for investment.

Some of the libraries utilized in this project:
•	react-table
•	react-currency-format
•	recharts
•	react-circular-progressbar
•	react-move

Custom Styling CSS file: StylingFile.css